refactor(content_ui): clarify shared type ownership
