+++ frontend/plugins/content_ui/src/modules/cms/shared/constants/languages.ts
