--- frontend/plugins/content_ui/src/modules/cms/types/index.ts
@@ -1,2 +1,0 @@
-export { IUser, ICategory, ITopic } from '../../shared/types';
-export type { IWebsite } from '../websites/types';
