--- frontend/plugins/content_ui/src/modules/shared/constants/index.ts
@@ -1,128 +1,0 @@
-import {
-  IconAlarm,
-  IconBriefcase,
-  IconCompass,
-  IconFlag,
-  IconPhoneCall,
-  IconBulb,
-  IconWorld,
-  IconHome,
-  IconPuzzle,
-  IconMedal,
-  IconThumbUp,
-  IconMusic,
-  IconPalette,
-  IconChartBar,
-  IconStar,
-  IconBrain,
-  IconSettings,
-  IconUsers,
-  IconPigMoney,
-  IconChartPie,
-  IconScale,
-  IconSpeakerphone,
-  IconTools,
-  IconUmbrella,
-  IconDiamond,
-  IconPencil,
-  IconDatabase,
-  IconVideo,
-  IconNote,
-  IconCamera,
-  IconArrowsShuffle,
-  IconHourglass,
-  IconMail,
-  IconGraph,
-  IconMessageDots,
-  IconMessage,
-  IconMoodSmile,
-  IconClock,
-  IconTrees,
-  IconGift,
-  IconMoon,
-  IconBook,
-  IconClipboard,
-  IconDeviceDesktop,
-  IconClipboardCopy,
-  IconSunset,
-  IconHeart,
-  IconFolder,
-} from '@tabler/icons-react';
-
-export const ICONS = [
-  { value: 'alarm', label: 'alarm', icon: IconAlarm },
-  { value: 'briefcase', label: 'briefcase', icon: IconBriefcase },
-  { value: 'earthgrid', label: 'earthgrid', icon: IconWorld },
-  { value: 'compass', label: 'compass', icon: IconCompass },
-  { value: 'idea', label: 'idea', icon: IconBulb },
-  { value: 'diamond', label: 'diamond', icon: IconDiamond },
-  { value: 'piggybank', label: 'piggybank', icon: IconPigMoney },
-  { value: 'piechart', label: 'piechart', icon: IconChartPie },
-  { value: 'scale', label: 'scale', icon: IconScale },
-  { value: 'megaphone', label: 'megaphone', icon: IconSpeakerphone },
-  { value: 'tools', label: 'tools', icon: IconTools },
-  { value: 'umbrella', label: 'umbrella', icon: IconUmbrella },
-  { value: 'bar-chart', label: 'bar chart', icon: IconChartBar },
-  { value: 'star', label: 'star', icon: IconStar },
-  { value: 'head-1', label: 'brain', icon: IconBrain },
-  { value: 'settings', label: 'settings', icon: IconSettings },
-  { value: 'users', label: 'users', icon: IconUsers },
-  { value: 'paintpalette', label: 'paint palette', icon: IconPalette },
-  { value: 'flag', label: 'flag', icon: IconFlag },
-  { value: 'phone-call', label: 'phone call', icon: IconPhoneCall },
-
-  { value: 'home', label: 'home', icon: IconHome },
-  { value: 'puzzle', label: 'puzzle', icon: IconPuzzle },
-  { value: 'medal', label: 'medal', icon: IconMedal },
-  { value: 'like', label: 'like', icon: IconThumbUp },
-  { value: 'book', label: 'book', icon: IconBook },
-  { value: 'clipboard', label: 'clipboard', icon: IconClipboard },
-  { value: 'computer', label: 'computer', icon: IconDeviceDesktop },
-  { value: 'paste', label: 'paste', icon: IconClipboardCopy },
-  { value: 'folder-1', label: 'folder', icon: IconFolder },
-  { value: 'sunset', label: 'sunset', icon: IconSunset },
-  { value: 'heart-alt', label: 'heart', icon: IconHeart },
-  { value: 'music-1', label: 'music', icon: IconMusic },
-  { value: 'pencil', label: 'pencil', icon: IconPencil },
-  { value: 'database', label: 'database', icon: IconDatabase },
-  { value: 'videocamera', label: 'video', icon: IconVideo },
-  { value: 'clipboard-1', label: 'note', icon: IconNote },
-  { value: 'camera', label: 'camera', icon: IconCamera },
-  { value: 'shuffle-1', label: 'shuffle', icon: IconArrowsShuffle },
-  { value: 'hourglass', label: 'hourglass', icon: IconHourglass },
-  { value: 'envelope-alt', label: 'envelope-alt', icon: IconMail },
-  { value: 'graph-bar', label: 'graph bar', icon: IconGraph },
-  { value: 'comment-alt-message', label: 'comment', icon: IconMessageDots },
-  { value: 'chat', label: 'chat', icon: IconMessage },
-  { value: 'smile', label: 'smile', icon: IconMoodSmile },
-  { value: 'wallclock', label: 'clock', icon: IconClock },
-  { value: 'trees', label: 'trees', icon: IconTrees },
-  { value: 'gift', label: 'gift', icon: IconGift },
-  { value: 'moon-1', label: 'night', icon: IconMoon },
-];
-
-export const REACTIONS = [
-  { value: 'https://erxes.s3.amazonaws.com/icons/sad.svg', label: 'Sad' },
-  {
-    value: 'https://erxes.s3.amazonaws.com/icons/neutral.svg',
-    label: 'Neutral',
-  },
-  {
-    value: 'https://erxes.s3.amazonaws.com/icons/grinning.svg',
-    label: 'Happy',
-  },
-  { value: 'https://erxes.s3.amazonaws.com/icons/like.svg', label: 'Like' },
-  {
-    value: 'https://erxes.s3.amazonaws.com/icons/dislike.svg',
-    label: 'Dislike',
-  },
-];
-
-export const LANGUAGES = [
-  { value: 'en', label: 'English' },
-  { value: 'mn', label: 'Mongolian' },
-  { value: 'zh', label: 'Chinese' },
-  { value: 'ja', label: 'Japanese' },
-  { value: 'ko', label: 'Korean' },
-  { value: 'ru', label: 'Russian' },
-];
